Top 50 Given-Name Characters in 6 Birth Cohorts

This dataset, top50char_year_df, is a data frame containing the top 50 most common Chinese characters used in given names across six birth cohorts: 1950, 1960, 1970, 1980, 1990, and 2000. It includes rankings and frequencies for all individuals, as well as separately for males and females. The dataset provides insights into naming character trends and gender differences over time.

Usage

data(top50char_year_df)

Format

A data frame with 50 observations and 37 variables:

top50

Ranking from 1 to 50 (numeric)

char.all.1950

Most common given-name character (all genders) in 1950 (character)

char.all.1960

Most common given-name character (all genders) in 1960 (character)

char.all.1970

Most common given-name character (all genders) in 1970 (character)

char.all.1980

Most common given-name character (all genders) in 1980 (character)

char.all.1990

Most common given-name character (all genders) in 1990 (character)

char.all.2000

Most common given-name character (all genders) in 2000 (character)

n.all.1950

Number of people with the character in 1950 (numeric)

n.all.1960

Number of people with the character in 1960 (numeric)

n.all.1970

Number of people with the character in 1970 (numeric)

n.all.1980

Number of people with the character in 1980 (numeric)

n.all.1990

Number of people with the character in 1990 (numeric)

n.all.2000

Number of people with the character in 2000 (numeric)

char.m.1950

Most common male given-name character in 1950 (character)

char.m.1960

Most common male given-name character in 1960 (character)

char.m.1970

Most common male given-name character in 1970 (character)

char.m.1980

Most common male given-name character in 1980 (character)

char.m.1990

Most common male given-name character in 1990 (character)

char.m.2000

Most common male given-name character in 2000 (character)

n.m.1950

Number of males with the character in 1950 (numeric)

n.m.1960

Number of males with the character in 1960 (numeric)

n.m.1970

Number of males with the character in 1970 (numeric)

n.m.1980

Number of males with the character in 1980 (numeric)

n.m.1990

Number of males with the character in 1990 (numeric)

n.m.2000

Number of males with the character in 2000 (numeric)

char.f.1950

Most common female given-name character in 1950 (character)

char.f.1960

Most common female given-name character in 1960 (character)

char.f.1970

Most common female given-name character in 1970 (character)

char.f.1980

Most common female given-name character in 1980 (character)

char.f.1990

Most common female given-name character in 1990 (character)

char.f.2000

Most common female given-name character in 2000 (character)

n.f.1950

Number of females with the character in 1950 (numeric)

n.f.1960

Number of females with the character in 1960 (numeric)

n.f.1970

Number of females with the character in 1970 (numeric)

n.f.1980

Number of females with the character in 1980 (numeric)

n.f.1990

Number of females with the character in 1990 (numeric)

n.f.2000

Number of females with the character in 2000 (numeric)

Source

Data taken from the ChineseNames package version 2023.8

Details

The dataset name has been kept as 'top50char_year_df' to avoid confusion with other datasets in the R ecosystem. This naming convention helps distinguish this dataset as part of the ChinAPIs package and assists users in identifying its specific characteristics. The suffix 'df' indicates that the dataset is a data frame. The original content has not been modified in any way.